get rid of ExtendedOWSConstants
