+++ core/src/main/java/org/n52/iceland/binding/AbstractXmlBinding.java
@@ -39,8 +39,8 @@ import org.n52.iceland.coding.decode.XmlStringOperationDecoderKey;
 import org.n52.iceland.exception.CodedException;
 import org.n52.iceland.exception.ows.NoApplicableCodeException;
 import org.n52.iceland.exception.ows.OwsExceptionReport;
-import org.n52.iceland.ogc.ows.ExtendedOWSConstants;
-import org.n52.iceland.ogc.ows.ExtendedOWSConstants.RequestParams;
+import org.n52.iceland.ogc.ows.OWSConstants;
+import org.n52.iceland.ogc.ows.OWSConstants.RequestParams;
 import org.n52.iceland.request.AbstractServiceRequest;
 import org.n52.iceland.request.Request;
 import org.n52.iceland.util.Constants;
@@ -85,13 +85,13 @@ public abstract class AbstractXmlBinding extends SimpleBinding {
                 opOrType = xmlStringKey.getOperation();
                 service = Optional.of(xmlStringKey.getService());
             }
-            if (ExtendedOWSConstants.Operations.GetCapabilities.toString().equalsIgnoreCase(opOrType)) {
+            if (OWSConstants.Operations.GetCapabilities.toString().equalsIgnoreCase(opOrType)) {
                 if (service.isPresent()) {
-                    throw new InvalidParameterValueException(ExtendedOWSConstants.GetCapabilitiesParams.service, service.get())
+                    throw new InvalidParameterValueException(OWSConstants.GetCapabilitiesParams.service, service.get())
                             .withMessage("The service '%s' is not supported.", service);
                 } else {
-                    throw new MissingParameterValueException(ExtendedOWSConstants.GetCapabilitiesParams.service)
-                            .withMessage("The parameter '%s' is missing.", ExtendedOWSConstants.GetCapabilitiesParams.service);
+                    throw new MissingParameterValueException(OWSConstants.GetCapabilitiesParams.service)
+                            .withMessage("The parameter '%s' is missing.", OWSConstants.GetCapabilitiesParams.service);
                 }
             } else {
                 throw new InvalidParameterValueException().withMessage(
@@ -113,7 +113,7 @@ public abstract class AbstractXmlBinding extends SimpleBinding {
             Document document = DocumentBuilderFactory.newInstance().newDocumentBuilder().parse(stream);
             Element element = document.getDocumentElement();
             element.normalize();
-            if (element.hasAttributes() && element.hasAttribute(ExtendedOWSConstants.RequestParams.service.name())) {
+            if (element.hasAttributes() && element.hasAttribute(OWSConstants.RequestParams.service.name())) {
                 OperationKey operationKey = getOperationKey(element);
                 XmlStringOperationDecoderKey decoderKey = new XmlStringOperationDecoderKey(operationKey, getDefaultContentType());
                 return decoderKey;
@@ -159,8 +159,8 @@ public abstract class AbstractXmlBinding extends SimpleBinding {
         String version = null;
         String operation = null;
         if (element.hasAttributes()) {
-            service = Strings.emptyToNull(element.getAttribute(ExtendedOWSConstants.RequestParams.service.name()));
-            version = Strings.emptyToNull(element.getAttribute(ExtendedOWSConstants.RequestParams.version.name()));
+            service = Strings.emptyToNull(element.getAttribute(OWSConstants.RequestParams.service.name()));
+            version = Strings.emptyToNull(element.getAttribute(OWSConstants.RequestParams.version.name()));
             if (!Strings.isNullOrEmpty(service)) {
                 String nodeName = element.getNodeName();
                 operation = nodeName.substring(nodeName.indexOf(Constants.COLON_STRING) + 1);
