+++ core/src/main/java/org/n52/iceland/exception/ows/concrete/InvalidServiceParameterException.java
@@ -17,7 +17,7 @@
 package org.n52.iceland.exception.ows.concrete;
 
 import org.n52.iceland.exception.ows.InvalidParameterValueException;
-import org.n52.iceland.ogc.ows.ExtendedOWSConstants;
+import org.n52.iceland.ogc.ows.OWSConstants;
 
 /**
  * @since 1.0.0
@@ -27,8 +27,8 @@ public class InvalidServiceParameterException extends InvalidParameterValueExcep
     private static final long serialVersionUID = 4979630437608155123L;
 
     public InvalidServiceParameterException(String value) {
-        at(ExtendedOWSConstants.GetCapabilitiesParams.service);
+        at(OWSConstants.GetCapabilitiesParams.service);
         withMessage("The value of the mandatory parameter '%s' is invalid. Delivered value was: %s",
-                ExtendedOWSConstants.GetCapabilitiesParams.service, value);
+                OWSConstants.GetCapabilitiesParams.service, value);
     }
 }
